requirements.txt
PyPDF2==3.0.1
PyMuPDF==1.23.7
google-generativeai==0.3.2
Pillow==10.2.0
httpx==0.27.0
python-dotenv==1.0.1 